requirements.txt
customtkinter>=5.2.0
yt-dlp>=2023.12.30
Pillow>=10.0.0
requests>=2.31.0
fontawesome-free>=5.15.4 